+++ objectified-ui/public/WHATS_NEW.md
@@ -7,6 +7,7 @@ We continue to improve the platform based on your feedback with improvements and
 ## New Features and Improvements
 
 - Moved import to its own feature, completely rewrote the logic
+- Added support for missing patternProperties in the property editor
 
 View our YouTube channel [here](https://www.youtube.com/@objectifieddev) for detailed tutorials and walkthroughs!